python-app/requirements.txt
flask==2.3.3
requests==2.28.2
pyyaml==5.4.1
cryptography==38.0.4
